chore: move iterables into their own folder
