--- src/scripts/utils/index.ts
@@ -1,30 +1,0 @@
-export * from './allEntries'
-export * from './average'
-export * from './chunk'
-export * from './entries'
-export * from './every'
-export * from './filter'
-export * from './first'
-export * from './fromEntries'
-export * from './gen'
-export * from './isIterable'
-export * from './Iter'
-export * from './join'
-export * from './last'
-export * from './map'
-export * from './partition'
-export * from './pipe'
-export * from './pluck'
-export * from './product'
-export * from './range'
-export * from './search'
-export * from './skip'
-export * from './skipWhile'
-export * from './some'
-export * from './sum'
-export * from './take'
-export * from './takeWhile'
-export * from './until'
-export * from './unzip'
-export * from './windows'
-export * from './zip'
