+++ src/scripts/utils/iterables/allEntries.ts
