+++ src/scripts/utils/iterables/average.ts
