+++ src/scripts/utils/iterables/entries.ts
