+++ src/scripts/utils/iterables/every.ts
