+++ src/scripts/utils/iterables/first.ts
