+++ src/scripts/utils/iterables/flat.ts
