+++ src/scripts/utils/iterables/fromEntries.ts
