+++ src/scripts/utils/iterables/gen.ts
