+++ src/scripts/utils/iterables/isIterable.ts
