+++ src/scripts/utils/iterables/last.ts
