+++ src/scripts/utils/iterables/map.ts
