+++ src/scripts/utils/iterables/partition.ts
