+++ src/scripts/utils/iterables/product.ts
