+++ src/scripts/utils/iterables/range.ts
