+++ src/scripts/utils/iterables/search.ts
