+++ src/scripts/utils/iterables/skip.ts
