+++ src/scripts/utils/iterables/skipWhile.ts
