+++ src/scripts/utils/iterables/some.ts
