+++ src/scripts/utils/iterables/sum.ts
