+++ src/scripts/utils/iterables/take.ts
