+++ src/scripts/utils/iterables/takeWhile.ts
