+++ src/scripts/utils/iterables/until.ts
