+++ src/scripts/utils/iterables/windows.ts
