+++ src/scripts/utils/pluckId.ts
@@ -1,3 +1,3 @@
-import { pluck } from "./pluck";
+import { pluck } from "./iterables/pluck";
 
 export const pluckId = pluck('id')
